Fix(pages): transfer the global styles from tha app to the main root route
