+++ src/data/product.json
@@ -2,12 +2,12 @@
   {
     "productName": "YX1 EARPHONES",
     "productThumbnail": "home/desktop/image-earphones-yx1.jpg",
-    "productURL": "/category-Y-headphones"
+    "productURL": "/category-Y-earphones"
   },
   {
     "productName": "ZX7 SPEAKER",
     "productThumbnail": "/local/desktop/zx7-speaker-bg.png",
-    "productURL": "/category-7-speakers"
+    "productURL": "/category-7-headphones"
   },
   {
     "productName": "ZX9 SPEAKER",
